Moving Submission form to its own componenet
