+++ frontend/src/components/forms/FormProps.interface.ts
@@ -1,0 +1,7 @@
+import { ArticleInterface } from "@/utils/article.interface";
+import { FormEvent } from "react";
+
+export interface FormProps {
+	setBody: (val: ArticleInterface[]) => void;
+	submitArticle: (e: FormEvent<HTMLFormElement>) => void;
+}
